backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.30.0
fastapi-sso==0.10.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.27.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
alembic>=1.13.0
